+++ R/article.R
@@ -168,6 +168,22 @@ mnras_article <- function(..., keep_tex = TRUE, fig_caption = TRUE) {
   )
 }
 
+#' @section \code{oup_article}: Format for creating submissions to many Oxford University Press
+#'   journals. Adapted from
+#'   \url{https://academic.oup.com/journals/pages/authors/preparing_your_manuscript}
+#'   and \url{https://academic.oup.com/icesjms/pages/General_Instructions}.
+#' @export
+#' @rdname article
+oup_article <- function(
+  ..., citation_package = 'natbib', keep_tex = TRUE,
+  md_extensions = c("-autolink_bare_uris")
+) {
+  pdf_document_format(
+    "oup_article", citation_package = citation_package,
+    keep_tex = keep_tex, md_extensions = md_extensions, ...
+  )
+}
+
 #' @section \code{peerj_article}: Format for creating submissions to The PeerJ
 #'   Journal. This was adapted from the
 #'   \href{https://www.overleaf.com/latex/templates/latex-template-for-peerj-journal-and-pre-print-submissions/ptdwfrqxqzbn}{PeerJ